Remove OrdersTableState and apply style to owner address
