+++ src/apps/explorer/components/OrdersTableWidget/context/OrdersTableContext.tsx
@@ -2,16 +2,10 @@ import React from 'react'
 
 import { Order } from 'api/operator'
 
-export enum OrdersTableState {
-  Loading,
-  Loaded,
-  Error,
-}
-
 interface CommonState {
   orders: Order[]
   error: string
-  kind: OrdersTableState
+  isFirstLoading: boolean
 }
 
 export const OrdersTableContext = React.createContext({} as CommonState)
